move and fix module imports
