+++ simprints/src/main/java/com/simprints/simprints/Constants.kt
@@ -1,0 +1,7 @@
+package com.simprints.simprints
+
+object Constants {
+    //SIMPRINTS
+    const val SIMPRINTS_NAMESPACE: String = "simprints"
+    const val SIMPRINTS_GUID: String = "simprintsGuid"
+}
