+++ simprints/src/main/java/com/simprints/simprints/SimprintsBiometricSearchResult.kt
@@ -1,4 +1,4 @@
-package org.dhis2.commons.simprints
+package com.simprints.simprints
 
 // The final matches by the Simprints biometric identification-based (one-to-many) search
 data class SimprintsBiometricSearchResult(
